requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Session==0.5.0
Werkzeug==2.3.7
email-validator==2.0.0
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
youtube-transcript-api==0.6.1
PyPDF2==3.0.1
